+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Connect tabs cycle with `Ctrl+Tab`. `Ctrl+Shift+Tab` no longer switches them.
+
 ## [1.0.1] - 2026-09-24
 
 ### Fixed